vector_store/requirements.txt
pyyaml>=6.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.0
tqdm>=4.66.0

